@@ -7,7 +7,7 @@ name = "qwikcrud"
 dynamic = ["version"]
 description = "An AI-powered command-line tool that generates RESTful APIs and admin interfaces based on user prompts"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = "Apache-2.0"
 keywords = ["ai", "chatgpt", "fastapi", "sqlalchemy", "starlette-admin", "crud"]
 authors = [
@@ -16,7 +16,6 @@ authors = [
 classifiers = [
     "Development Status :: 4 - Beta",
     "Programming Language :: Python",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -65,7 +64,7 @@ lint = [
 ]
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 line-length = 120
 select = [
     "A",
